+++ src/components/AutocompleteInput/AutocompleteInput.stories.tsx
@@ -20,7 +20,7 @@ export default {
  *
  * @param props the story props
  */
-const Template: Story<AutocompleteInputProps> = (props) => {
+const Template: Story<AutocompleteInputProps<{ id: number; label: string }>> = (props) => {
   const options = Array(10)
     .fill(null)
     .map((_, i) => ({
@@ -34,7 +34,6 @@ const Template: Story<AutocompleteInputProps> = (props) => {
       options={options}
       itemToString={(item) => (item ? item.label : '')}
       filterItem={(item, inputValue) => item.label.toLowerCase().includes(inputValue.toLowerCase())}
-      initialSelectedItem={options[0]}
     />
   );
 };